requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Faker==28.4.1
helper==2.5.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pip==24.2
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
